requirements.txt
cryptography
pillow
pydub
wave
pytest
tk
